Fix bulk send
