+++ src/server/api/mutations/index.js
@@ -1,0 +1,15 @@
+import {
+  sendMessage
+} from "./sendMessage";
+import {
+  bulkSendMessages
+} from "./bulkSendMessages";
+import {
+  findNewCampaignContact
+} from "./findNewCampaignContact";
+
+export {
+  sendMessage,
+  bulkSendMessages,
+  findNewCampaignContact
+};
